# Conflicts:
#	core-bundle/src/Resources/contao/library/Contao/DcaLoader.php
#	manager-bundle/src/HttpKernel/ContaoKernel.php

+++ installation-bundle/src/Database/Version480Update.php
@@ -233,6 +233,8 @@ class Version480Update extends AbstractVersionUpdate
         // Since rememberme is broken in Contao 4.7 and there are no valid
         // cookies out there, we can simply drop the old table here and let the
         // install tool create the new one
-        $this->connection->query('DROP TABLE tl_remember_me');
+        if ($this->connection->getSchemaManager()->tablesExist(['tl_remember_me'])) {
+            $this->connection->query('DROP TABLE tl_remember_me');
+        }
     }
 }
